Rewrites EscapeAllHTMLTags() to properly parse tags.
As of now all tags that conform to TokenType in html package in
golang is escaped. This means ">" and "<" should no longer be
escaped if not part of a tag.

Also adds the missing unit test for the EscapeAllHTMLTags().

Reference: https://github.com/golang/net/blob/master/html/token.go#L18