Adds a list of HTML tags which are used for escaping.
Not all tags tokenized are escaped as of now, it is compared against
a list of known HTML tags, which are escaped. This means a nickname
like <bob> will no longer be escaped as &lt;bob&gt; but will be
shown as <bob> in notification.

The test has been updated to reflect this change.

Related to issue #508