muc-room-dialog-error-component - Uses a custom structure to pass the options to the initializer.
The idea behind this commit is to have a clean way of passing multiple parameters to the error dialog and to avoid knowing the order of them, and other things.

This is done to allow the passing of a "parent window" property in the next commit, to avoid the "dialog with no transient" warning.